requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
dnspython==1.16.0
Flask==1.0.2
flask-mongoengine==0.9.5
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-Testing==0.7.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
mongoengine==0.17.0
passlib==1.7.1
py3dns==3.2.0
pycparser==2.19
pyjson==1.3.0
PyJWT==1.4.2
pymongo==3.7.2
pyOpenSSL==19.0.0
pyrsistent==0.14.11
pytz==2019.1
six==1.12.0
Werkzeug==0.15.2
WTForms==2.2.1
